+++ lib/countries/country.rb
@@ -34,24 +34,6 @@ module ISO3166
     alias zip_format postal_code_format
     alias languages languages_official
 
-    def latitude_dec
-      if RUBY_VERSION =~ /^3\.\d\.\d/
-        warn "DEPRECATION WARNING: The Country#latitude_dec method has been deprecated and will be removed in 5.0. Please use Country#latitude instead.", uplevel: 1, category: :deprecated
-      else
-        warn "DEPRECATION WARNING: The Country#latitude_dec method has been deprecated and will be removed in 5.0. Please use Country#latitude instead.", uplevel: 1
-      end
-      latitude
-    end
-
-    def longitude_dec
-      if RUBY_VERSION =~ /^3\.\d\.\d/
-        warn "DEPRECATION WARNING: The Country#longitude_dec method has been deprecated and will be removed in 5.0. Please use Country#longitude instead.", uplevel: 1, category: :deprecated
-      else
-        warn "DEPRECATION WARNING: The Country#longitude_dec method has been deprecated and will be removed in 5.0. Please use Country#longitude instead.", uplevel: 1
-      end
-      longitude
-    end
-
     def ==(other)
       other.respond_to?(:alpha2) && other.alpha2 == alpha2
     end
@@ -137,28 +119,6 @@ module ISO3166
     # @return [String] The name for this Country, in this Country's locale.
     def local_name
       @local_name ||= local_names.first
-    end
-
-    # @return [String] This Country's ISO Short Name
-    # @deprecated Use {#iso_short_name} instead.
-    def name
-      if RUBY_VERSION =~ /^3\.\d\.\d/
-        warn "DEPRECATION WARNING: The Country#name method has been deprecated. Please use Country#iso_short_name instead or refer to the README file for more information on this change.", uplevel: 1, category: :deprecated
-      else
-        warn "DEPRECATION WARNING: The Country#name method has been deprecated. Please use Country#iso_short_name instead or refer to the README file for more information on this change.", uplevel: 1
-      end
-      iso_short_name
-    end
-
-    # @return [Array<String>] Array of unofficial, slang names or aliases for this Country
-    # @deprecated Use {#unofficial_names} instead.
-    def names
-      if RUBY_VERSION =~ /^3\.\d\.\d/
-        warn "DEPRECATION WARNING: The Country#names method has been deprecated. Please use Country#unofficial_names instead or refer to the README file for more information on this change.", uplevel: 1, category: :deprecated
-      else
-        warn "DEPRECATION WARNING: The Country#names method has been deprecated. Please use Country#unofficial_names instead or refer to the README file for more information on this change.", uplevel: 1
-      end
-      unofficial_names
     end
 
     # @!attribute alpha2