+++ spec/country_spec.rb
@@ -48,7 +48,7 @@ describe ISO3166::Country do
   end
 
   it 'should return alternate names' do
-    expect(country.names).to eq(['United States', 'USA', 'Murica',
+    expect(country.unofficial_names).to eq(['United States', 'USA', 'Murica',
                                  'Vereinigte Staaten von Amerika', 'États-Unis',
                                  'Estados Unidos', 'アメリカ合衆国', 'Verenigde Staten'])
   end
@@ -717,6 +717,49 @@ describe ISO3166::Country do
       end
     end
 
+    describe '#find_country_by_any_name' do
+      context 'when search name found' do
+        let(:uk) { ISO3166::Country.find_country_by_any_name('United Kingdom') }
+
+        it 'should be a country instance' do
+          expect(uk).to be_a(ISO3166::Country)
+          expect(uk.alpha2).to eq('GB')
+        end
+      end
+
+      context 'when search lowercase name found' do
+        let(:uk) { ISO3166::Country.find_country_by_any_name('united kingdom') }
+
+        it 'should be a country instance' do
+          expect(uk).to be_a(ISO3166::Country)
+          expect(uk.alpha2).to eq('GB')
+        end
+      end
+
+      context 'when the search term contains comma' do
+        let(:korea) { ISO3166::Country.find_country_by_any_name('Korea, Republic of') }
+
+        it 'should be a country instance' do
+          expect(korea).to be_a(ISO3166::Country)
+          expect(korea.alpha2).to eq('KR')
+        end
+      end
+
+      context 'when search translation found' do
+        before do
+          ISO3166.configure do |config|
+            config.locales = [:bs]
+          end
+        end
+        let(:uk) { ISO3166::Country.find_country_by_any_name('Velika Britanija') }
+
+        it 'should be a country instance' do
+          expect(uk).to be_a(ISO3166::Country)
+          expect(uk.alpha2).to eq('GB')
+        end
+      end
+    end
+
     context 'regression test for #388' do
       let(:no_country) { ISO3166::Country.find_country_by_translated_names(nil) }
 
@@ -733,6 +776,7 @@ describe ISO3166::Country do
         expect { uk }.to raise_error(RuntimeError)
       end
     end
+
     context 'when search name not found' do
       let(:bogus) { ISO3166::Country.find_country_by_unofficial_names('Does not exist') }
 
@@ -853,7 +897,7 @@ describe ISO3166::Country do
   end
 
   describe 'names in Data' do
-    it 'should be unique (to allow .find_by_name work properly)' do
+    it 'should be unique (to allow .find_by_any_name work properly)' do
       names = ISO3166::Data.cache.map do |_k, v|
         [v['iso_short_name'], v['unofficial_names']].flatten.uniq
       end.flatten